requirements.txt
requests
pyyaml
absl-py
influxdb
